+++ CHANGELOG.md
@@ -14,6 +14,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ### fixed
 
+- Keep sanger order + verification when updating/reloading variants
+
 
 ## [4.7.3]